backend/requirements.txt
fastapi
uvicorn
python-dotenv
requests
openai
huggingface_hub
pillow
